backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
langchain>=0.3.26
langchain-chroma>=0.2.4
langchain-community>=0.3.26
langchain-openai>=0.3.25
langgraph>=0.4.9
pandas>=2.0.0
python-dotenv>=1.1.1
watchdog>=4.0.0
pydantic>=2.0.0 